Rename CartoDB importer to "Uploader"
This is slightly more representative of what we are doing here. While we
are definitely importing to cartodb, we are really just using the import
api as a tool to upload to CartoDB so that we can hack around with and
merge our split up shapefiles. The uploader and merger will be handled
by a parent import class, which I think is more aptly named.