Refactors stats controller code in to presenters
Previously the stats controllers had quite a lot of say in how the stats
were presented (e.g. deciding whether or not to round the values, etc.).
This responsibility has moved to the `StatisticPresenter`.

StatisticPresenter proxies methods to the model or model statistic, and
presents the data differently depending on what was requested. For
example, the global pa coverage percentage is rounded to one significant
digit.