Search reindexer bulk imports models to Elastic Search
Indexing with Elastic Search is different: it requires the attributes of
the searchable models to be saved in to the Elastic Search index, a la
mongodb. This is pretty slow, so here we use the ES _bulk import API by
building JSON structures for the given models and then sending them to
the API.

Currently batched to 1000 records per API call to hopefully reduce any
memory usage issues.