Adds ability to boost search results via the query
In this case, we are boosting the score for matching countries and
regions so that they appear higher in the results.

Scoring is skewed towards PAs, despite countries and regions having
better matched names, because the PAs contain more content (sub
location, designation, nested country, etc.) and is thus more likely to
match the given query.