Large refactoring of Javascripts
Previously all JS would be loaded on all pages. This is now restricted
on a controller basis, to make deciding what map to render easier: the
PA page uses the PA JS controller, etc.

Alongside this, the PA Map has been refactored out in to a generic map,
with the PA map handling only the rendering of the CartoDB tiles. This
way, the soon-to-exist search results map can separately hold its own
bits of logic, thus preventing a giant map class to solve all problems.