Corrects nginx API passthrough
The ^~ operator in nginx matches routes that begin with the given
string, and then stops searching. This allows us to use the order of
location directives to decide what to do. In this case we give priority
to legacy API requests, in order of specificity, so that they are not
accidentally passed to the new app (which does not support these
routes).